add some tests for getWordsFromMessages
