+++ backend/utils/filterSlackResponse.js
@@ -42,6 +42,7 @@ const AddThreadToParent = (thread, messages, parentIndex) => {
 }
 
 const GetWordsFromMessages = (messages) => {
+  console.log(JSON.stringify(messages))
   const result = []
   const temp_word_obj = {}
